requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.3.0
PyJWT==2.8.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pytesseract==0.3.10
python-dateutil==2.9.0.post0
pytz==2024.1
pyupbit==0.2.33
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
tesseract==0.1.3
tzdata==2024.1
urllib3==2.2.1
websockets==12.0
